Replacing start up code snippets with FreeBSD version.
The original Bare Metal Container code borrowed some functions and
macros from Linux, which uses a GPLv2, or copy-left, license. The
problem arises because the payload application ELF is technically
linked to this GPLv2 code, so an argument could be made that the
source code of that payload application (i.e., end customer's
secret sauce) would also have to be distributed using the GPLv2
license. A counter argument could be made that since the  payload
application is turned into a binary before being linked the guest
ready application would not be a derivative product, but it was
decided that it would just be better all around to replace the
GPLv2 code with FreeBSD equivalent.