Added more flexible MMU functionality.
This commits adds the 'map_memory' function, which allows updates
of the MMU down to L3, 4KB page. This update also adds that
along with the Xen console print function to table at a fixed
so the payload application can lookup and use those functions,
which is useful for mapping new pages (like for PV console) or
having a way to print messages if PV console or UART passthrough
isn't working.